chore: Move IntersectionObserver mock to pipeline src
