+++ packages/pipeline/src/Stage.spec.tsx
@@ -1,4 +1,4 @@
-import { mockIsIntersecting } from '../../../__mocks__/IntersectionObserver';
+import { mockIsIntersecting } from './mocks/IntersectionObserver';
 
 import React from 'react';
 import { typeIs } from '@leafygreen-ui/lib';
