+++ packages/pipeline/src/mocks/IntersectionObserver.ts
@@ -1,20 +1,24 @@
 import { act } from 'react-dom/test-utils';
+import { IntersectionOptions } from 'react-intersection-observer';
 
 const observerMap = new Map();
 const instanceMap = new Map();
 
-function IntersectionObserver(cb, options) {
+function IntersectionObserver(
+  cb: IntersectionObserverCallback,
+  options: IntersectionOptions,
+) {
   const instance = {
     thresholds: Array.isArray(options.threshold)
       ? options.threshold
       : [options.threshold],
     root: options.root,
     rootMargin: options.rootMargin,
-    observe: element => {
+    observe: (element: Element) => {
       instanceMap.set(element, instance);
       observerMap.set(element, cb);
     },
-    unobserve: element => {
+    unobserve: (element: Element) => {
       instanceMap.delete(element);
       observerMap.delete(element);
     },
@@ -25,12 +29,21 @@ function IntersectionObserver(cb, options) {
 }
 
 /**
- * Set the `isIntersecting` for the IntersectionObserver of a specific element.
- *
- * @param {HTMLElement} element
- * @param {boolean} isIntersecting
+ * Set the `isIntersecting` on all current IntersectionObserver instances
+ * @param isIntersecting {boolean}
  */
-function mockIsIntersecting(element, isIntersecting) {
+function mockAllIsIntersecting(isIntersecting: boolean) {
+  observerMap.forEach((onChange, element) => {
+    mockIsIntersecting(element, isIntersecting);
+  });
+}
+
+/**
+ * Set the `isIntersecting` for the IntersectionObserver of a specific element.
+ * @param element {Element}
+ * @param isIntersecting {boolean}
+ */
+function mockIsIntersecting(element: Element, isIntersecting: boolean) {
   const cb = observerMap.get(element);
 
   if (cb) {
@@ -41,12 +54,8 @@ function mockIsIntersecting(element, isIntersecting) {
         intersectionRatio: isIntersecting ? 1 : 0,
       },
     ];
-
-    if (act) {
-      act(() => cb(entry));
-    } else {
-      cb(entry); // eslint-disable-line callback-return
-    }
+    if (act) act(() => cb(entry));
+    else cb(entry);
   } else {
     throw new Error(
       'No IntersectionObserver instance found for element. Is it still mounted in the DOM?',
@@ -55,25 +64,13 @@ function mockIsIntersecting(element, isIntersecting) {
 }
 
 /**
- * Set the `isIntersecting` on all current IntersectionObserver instances
- *
- * @param {Boolean} isIntersecting
- */
-function mockAllIsIntersecting(isIntersecting) {
-  observerMap.forEach((onChange, element) => {
-    mockIsIntersecting(element, isIntersecting);
-  });
-}
-
-/**
  * Call the `intersectionMockInstance` method with an element, to get the (mocked)
  * `IntersectionObserver` instance. You can use this to spy on the `observe` and
  * `unobserve` methods.
- *
- * @param {HTMLElement} element
- * @return {Object} IntersectionObserver
+ * @param element {Element}
+ * @return IntersectionObserver
  */
-function intersectionMockInstance(element) {
+function intersectionMockInstance(element: Element): IntersectionObserver {
   return instanceMap.get(element);
 }
